Move code into src folder
