+++ README.md
@@ -5,5 +5,17 @@ This project applies Reinforcement Learning (RL) algorithms to solve Sokoban puz
 - Monte Carlo Policy Optimization (Every Visit). Usage:
 
 ```bash
-python montecarlo.py <puzzle_file> <number_of_episodes> <discount_factor> <exploration_rate>
+python .\montecarlo.py <puzzle_file> <number_of_episodes> <discount_factor> <exploration_rate>
+```
+
+Example:
+
+```bash
+python .\montecarlo.py .\puzzles\easy.txt 100000 0.85 0.9
+```
+
+To play the game yourself, run:
+
+```bash
+python .\humanPlayer.py <puzzle_file>
 ```
